chore: interface 파일 추가
