+++ src/components/common/Chart/LineChart/LineChart.utils.ts
@@ -1,29 +1,6 @@
 import { getMonthName } from "../chart.utils";
 
-export interface ILineChart {
-  tracks: ILineChartList[]
-}
-
-interface ILineChartList {
-  id: number,
-  name: string,
-  enName: string,
-  monthlyTrend: ILineChartItem[]
-}
-
-interface ILineChartItem {
-  month: number,
-  settlement: number,
-  revenue: number,
-}
-
-export interface IMappedChartData {
-  id: number,
-  data: {
-    x: string,
-    y: number
-  }[]
-}
+import { IMappedChartData, ILineChart } from "./LineChart.types";
 
 export const getMaxValueInLineChart = (chartList: IMappedChartData[]) => {
   let maxValue = 0;
